Split Version class to move all about supported versions into a specific QtiVersion, keeping only semantic versioning matters in the actual Version class.
