+++ src/qtism/common/utils/versions/QtiVersion300.php
@@ -21,14 +21,12 @@
  */
 namespace qtism\common\utils\versions;
 
-use qtism\common\utils\Version;
-
 /**
  * QTI version 3.0.0
  * 
  * This class is a draft since QTI 3.0 hasn't been published yet.
  */
-class QtiVersion300 extends Version
+class QtiVersion300 extends QtiVersion
 {
     const XMLNS = 'http://www.imsglobal.org/xsd/imsaqti_item_v1p0';
